backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pymongo
motor
python-dotenv==1.0.1
google-generativeai==0.8.3
sentence-transformers==3.3.1
faiss-cpu>=1.9.0.post1
numpy>=1.26.4
pydantic==2.10.3
python-multipart==0.0.18
httpx==0.27.2
tf-keras
